refactor(repository,controllers): fix UserChatsController & clean up repo beans
* remove redundant third repository interface;
* make UserChatsController working;
* add stubs to services to support new feature;
* move application tests to appropreate package;
