+++ src/main/java/com/github/alebabai/tg2vk/repository/UserRestRepository.java
@@ -5,12 +5,10 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.security.access.prepost.PostAuthorize;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.stereotype.Repository;
 
-@Repository("userRestRepository")
 @PreAuthorize("hasRole('ROLE_ADMIN')")
 @RepositoryRestResource(path = "users", collectionResourceRel = "users", itemResourceRel = "user")
-public interface UserRestRepository extends UserBaseRepository {
+public interface UserRestRepository extends UserRepository {
     @PreAuthorize("hasAnyRole('ROLE_ADMIN','ROLE_USER') and #user?.tgId == principal")
     @Override
     <S extends User> S save(@Param("user") S user);
