+++ src/main/java/com/github/alebabai/tg2vk/service/VkService.java
@@ -1,10 +1,12 @@
 package com.github.alebabai.tg2vk.service;
 
+import com.github.alebabai.tg2vk.domain.Chat;
+import com.github.alebabai.tg2vk.domain.User;
 import com.vk.api.sdk.client.actors.Actor;
 import com.vk.api.sdk.client.actors.UserActor;
 import com.vk.api.sdk.objects.messages.Message;
-import com.vk.api.sdk.objects.users.User;
 
+import java.util.Collection;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.BiConsumer;
@@ -16,5 +18,7 @@ public interface VkService {
 
     String getAuthorizeUrl(String redirectUrl, String... scopes);
 
-    CompletableFuture<Integer> fetchMessages(Actor actor, BiConsumer<User, Message> consumer);
+    CompletableFuture<Integer> fetchMessages(Actor actor, BiConsumer<com.vk.api.sdk.objects.users.User, Message> consumer);
+
+    Collection<Chat> getChats(User user);
 }
